requirements.txt
#packages 
langchain
langchain-openai
openai
pinecone
python-dotenv
langchain-community
langchain-pinecone
